sat_tools/requirements.txt
# SAT Tools Python Dependencies

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# HTTP Client
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# JSON Schema Validation (optional)
jsonschema>=4.20.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.7.0

# Linting
ruff>=0.1.0
